+++ src/redsun/containers/_config.py
@@ -15,3 +15,4 @@ class AppConfig(RedSunConfig, total=False):
     devices: NotRequired[dict[str, Any]]
     presenters: NotRequired[dict[str, Any]]
     views: NotRequired[dict[str, Any]]
+    wiring: NotRequired[list[dict[str, str]]]